--- microk8s-webconsole/commons/types/graphql.schema.ts
@@ -1,10 +1,0 @@
-export class Addon {
-    name: string;
-    enabled: boolean;
-}
-
-export abstract class IQuery {
-    abstract getAddons(): Addon[] | Promise<Addon[]>;
-
-    abstract temp__(): boolean | Promise<boolean>;
-}
