+++ microk8s-webconsole/server/src/microk8s/addons/addons.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from '@nestjs/common';
 import { safeLoad } from 'js-yaml';
-import { Addon } from '@common/models/addon.interface';
+import { Addon } from '../../../../common/models/addon.interface';
 import { from } from 'rxjs';
 import { map } from 'rxjs/operators';
-import {ShellService} from '../core/services/shell/shell.service';
+import {ShellService} from '../../core/services/shell/shell.service';
 
 @Injectable()
 export class AddonsService {
