+++ microk8s-webconsole/server/src/microk8s/microk8s.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { AddonsModule } from './addons/addons.module';
+
+@Module({
+  imports: [AddonsModule],
+  exports: [AddonsModule],
+})
+export class Microk8sModule {}
