+++ microk8s-webconsole/server/src/snap/snap.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { SnapService } from './snap.service';
+import { SnapController } from './snap.controller';
+
+@Module({
+  providers: [SnapService],
+  controllers: [SnapController],
+})
+export class SnapModule {}
